Refactor: Make `RemoteConfigStoreMock` accessible from other tests
